CQT Experiment: Tossed in a branch for later
This experiment with Constant Q Transform was interesting, and it's nice
to see that it actually functions, but it's just too damn slow to really
be useful for visualization purposes. It uses nearly a full core on an
M1 processor, and I'd hate to see what it does on an Intel machine.
Stashing this in a branch and discarding it from the main tree, in case
somebody finds this code useful some day.

Signed-off-by: Christopher Snowhill <kode54@gmail.com>