README: Add directions for Git hooks path
Since we use a non-default path in the repository to store shared hooks,
it needs to be configured on a freshly cloned repository before it can
be usable.

Yeah, I forgot to follow my own directions before working on the project
for quite some time. The error crept in at the following commit:

2aa3ddd545f376fb6d1c6932cafc5e701a09a47e - Icon replacement, and team ID

Signed-off-by: Christopher Snowhill <kode54@gmail.com>