Makes volume logarithmic when limited to 100%
Modifies Helper.m, Helper.h, and VolumeSlider.m so when volume is limited to 100%, the volume slider is logarithmic with an X^2 approximation.

Also makes sure the value shown on the volume tooltip is not logarithmic when limited to 100%.

Makes some variables in VolumeSlider.m and parameters in Helper.h/Helper.m const as well.