Visualization: Correct possible race condition
It's safer to ignore an out of range latency value before it's actually
assigned, rather than letting it slip through before gating it with the
control variable. So, assign the ignore latency state before assigning
latency itself, since neither side is synchronization locked, because it
used to just be one variable, which was fairly safe to leave unsynced,
but now it's two.

Signed-off-by: Christopher Snowhill <kode54@gmail.com>