Playback Control: Workaround for fade in/out
Dirty hack to work around the fact that fade out restores the volume
level after it completes, which would otherwise work, if not for the
fade-based pausing and unpausing also implemented! So pause once, then
wait out the pause fade interval, then restore the volume level.

No hack is needed for fading back in, because the two fade-ins will
simply overlap, and not cause any glitch. Nice!

Fixes #440

Signed-off-by: Christopher Snowhill <kode54@gmail.com>