FFmpeg: Update to version 8.0
Update FFmpeg to version 8.0, carrying the same patches, but updating
one of them to reflect changes to the original library which temporarily
broke it.

Looks like we already caught a change to interfaces which would have
broken compilation with 8.0, by the fact it was declared a deprecated
function call in earlier releases.

Still carrying with fdk-aac. Also replaced yasm with nasm, and newly
imported pkgconf, which was apparently already manually installed on the
previous build systems, and therefore was missed as a build dependency
for the FFmpeg scripts to detect our libfdk-aac.

Also eliminated a switch to disable building postproc, as that's
apparently been split out of the FFmpeg tree.

Signed-off-by: Christopher Snowhill <kode54@gmail.com>