Plugin Controller: Fix mutex delete race condition
There was a potential race condition where the mutex object would be
deleted before the Plugin Controller was, so the cache thread crashes
trying to lock a deleted mutex object.

Change it so it allocates the mutex dynamically on the heap before
starting the thread, and destroys it after stopping the thread.

Also be sure to NULL the pointers, so any possible racy crash will
result from NULL pointer dereference instead of accessing freed memory.

Signed-off-by: Christopher Snowhill <kode54@gmail.com>