Output: Don't dispatch background events on stop
When stopping the player, it should not be dispatching background events
onto the main thread which will touch the output. Or at least, if it has
already been dispatched, it should check if the player is in the middle
of stopping already before trying to call functions on the object.

Signed-off-by: Christopher Snowhill <kode54@gmail.com>