Jaytrax/Syntrax: Bring back decoder with changes
Significant changes have been made to the original decoder, including an
entirely unnecessary big endian support mode. The API was changed a bit
as well, and some further error checking was added. Also, the echo
buffers were extended significantly to accommodate higher sample rates,
and the calculation for wrapping the echo buffer pointer was moved to a
less trafficked spot in the player, to improve performance. Finally, the
resampler loop for sampled instruments was changed significantly, to
reduce the amount of copying performed during playback.

Oh yeah, and a copyright notice and license was added to the library. If
the original developer is not satisfied with the choice of license, they
may recommend something different. After all, this was done based on
code that bore no upstream license. It is assumed that changes from
downstream will be wanted upstream some day, perhaps. Or at least this
may become an authoritative upstream.

The library will also be re-uploaded to a Bitbucket repository.

Signed-off-by: Christopher Snowhill <kode54@gmail.com>