Visualization: Eliminate flickering problem
Also eliminate thread locking for FFT code, by requiring the caller to
supply a pointer which will be allocated with a state structure for the
FFT calculation code.

The flickering problem was caused by the visAudioTemp pointer being a
void type, so incrementing it on buffer wraps by bytes instead of floats
as it should have, which caused it to overwrite part of itself, and also
leave the majority of the buffer empty.

It was first attempted to change the pointer type to float*, but instead
we arrived at requiring the caller to supply a PCM buffer as well if FFT
is desired, so that took care of yet another long standing bug.

The bug would present as the visualization data being intermittently
silent every roughly 45 seconds or so, for a random number of frames
which happened to overlap the edge of the circular PCM buffer. Whew.

Signed-off-by: Christopher Snowhill <kode54@gmail.com>