Highly Complete: Move silence seconds declaration
Declare the silence seconds variable as a state member, so other
functions may use it. In this case, so that the buffer filling function
will at least try to fill up a full buffer if no total length has been
reported yet, as will happen when GSF is doing early initialization for
sample rate detection.

Signed-off-by: Christopher Snowhill <kode54@gmail.com>