FFmpeg: Signal tag updates for normal files
This signal should be reported if the FFmpeg input has reported any tag
information for seekable static files. Previously, this value may have
had some sort of meaning, but prior to this change, has not actually
done anything.

Also fixes #458 I think. Not sure if this change was 100% needed for
that particular bug.

Signed-off-by: Christopher Snowhill <kode54@gmail.com>