Update Sentry framework to version 9.1.0
And update the default settings, including removing the unnecessary PII
(IP address) collection, as we really only need an install ID to
differentiate unique installs and the crashes which are occurring there.

Technically the server still sees the IP address of the user connecting
to submit the crash reports and event logs, when enabled, but it doesn't
record that information from the session.

Signed-off-by: Christopher Snowhill <kode54@gmail.com>